requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile requirements.in
#
black==23.1.0
    # via -r requirements.in
click==8.1.3
    # via
    #   -r requirements.in
    #   black
mypy-extensions==0.4.3
    # via black
numpy==1.24.1
    # via -r requirements.in
packaging==23.0
    # via black
paho-mqtt==1.6.1
    # via -r requirements.in
pathspec==0.11.0
    # via black
platformdirs==2.6.2
    # via black
pydantic==1.10.4
    # via -r requirements.in
tomli==2.0.1
    # via black
typing-extensions==4.4.0
    # via pydantic
